refactor: extract levenshtein/fuzzyMatch to utils/search.js
